move folders into pkg folder
