+++ src/pkg/cnfuzz/persistence/cache.go
@@ -19,10 +19,10 @@ package persistence
 import (
 	"context"
 	"fmt"
-	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/health"
 	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/model"
 	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/persistence/in_memory"
 	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/persistence/redis"
+	"github.com/suecodelabs/cnfuzz/src/pkg/health"
 	"github.com/suecodelabs/cnfuzz/src/pkg/logger"
 )
