+++ src/pkg/cnfuzz/persistence/redis/container_image.go
@@ -18,8 +18,8 @@ package redis
 
 import (
 	"context"
-	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/health"
 	"github.com/suecodelabs/cnfuzz/src/pkg/cnfuzz/model"
+	"github.com/suecodelabs/cnfuzz/src/pkg/health"
 	"github.com/suecodelabs/cnfuzz/src/pkg/logger"
 	"time"
