+++ src/pkg/k8s/util/container_image.go
