+++ src/pkg/k8s/util/container_image_test.go
